해시태그 리스트를 String 으로 받아서 array 나 arrayList로 저장하는 법

+++ src/main/java/com/team2/docgram/dto/BoardDto.java
@@ -22,7 +22,7 @@ public class BoardDto {
 	private Timestamp date;
 	private Timestamp start_date;
 	private Timestamp end_date;
-	private int user;
+	private Long user;
 	private int file;
 	private int hashtag;
 	private UserDto userDetail;
